actions: Split out zerver.actions.realm_playgrounds.
(cherry picked from commit e887abcf41b969d7a29582c8550de217f7a6823b)
