+++ zerver/openapi/curl_param_value_generators.py
@@ -10,10 +10,10 @@ from typing import Any, Callable, Dict, List, Optional, Set, Tuple
 
 from django.utils.timezone import now as timezone_now
 
+from zerver.actions.realm_playgrounds import do_add_realm_playground
 from zerver.lib.actions import (
     do_add_linkifier,
     do_add_reaction,
-    do_add_realm_playground,
     do_create_user,
     update_user_presence,
 )
