+++ zerver/views/realm_playgrounds.py
@@ -4,8 +4,8 @@ from django.core.exceptions import ValidationError
 from django.http import HttpRequest, HttpResponse
 from django.utils.translation import gettext as _
 
+from zerver.actions.realm_playgrounds import do_add_realm_playground, do_remove_realm_playground
 from zerver.decorator import require_realm_admin
-from zerver.lib.actions import do_add_realm_playground, do_remove_realm_playground
 from zerver.lib.exceptions import JsonableError, ValidationFailureError
 from zerver.lib.request import REQ, has_request_variables
 from zerver.lib.response import json_success
